+++ wallets/leap-mobile/src/wallet-connect/registry.ts
@@ -44,7 +44,7 @@ export const LeapMobileInfo: Wallet = {
       appUrl: string,
       wcUri: string,
       os: OS | undefined,
-      name: string
+      _name: string
     ): string => {
       const plainAppUrl = appUrl.replaceAll('/', '').replaceAll(':', '');
       const encodedWcUrl = encodeURIComponent(wcUri);