+++ packages/cosmos-kit/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0](https://github.com/cosmology-tech/cosmos-kit/compare/cosmos-kit@2.0.5...cosmos-kit@2.1.0) (2023-07-28)
+
+**Note:** Version bump only for package cosmos-kit
+
+## [2.0.5](https://github.com/cosmology-tech/cosmos-kit/compare/cosmos-kit@2.0.4...cosmos-kit@2.0.5) (2023-07-28)
+
+**Note:** Version bump only for package cosmos-kit
+
+## [2.0.4](https://github.com/cosmology-tech/cosmos-kit/compare/cosmos-kit@0.12.3...cosmos-kit@2.0.4) (2023-07-27)
+
+**Note:** Version bump only for package cosmos-kit
+
 ## [0.12.3](https://github.com/cosmology-tech/cosmos-kit/compare/cosmos-kit@0.12.2...cosmos-kit@0.12.3) (2023-07-24)
 
 **Note:** Version bump only for package cosmos-kit