+++ packages/example/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.1](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.1.0...example@2.1.1) (2023-07-28)
+
+**Note:** Version bump only for package example
+
+# [2.1.0](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.0.5...example@2.1.0) (2023-07-28)
+
+**Note:** Version bump only for package example
+
+## [2.0.5](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.0.4...example@2.0.5) (2023-07-28)
+
+**Note:** Version bump only for package example
+
+## [2.0.4](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.0.3...example@2.0.4) (2023-07-27)
+
+**Note:** Version bump only for package example
+
 ## [2.0.3](https://github.com/cosmology-tech/cosmos-kit/compare/example@2.0.2...example@2.0.3) (2023-07-24)
 
 **Note:** Version bump only for package example