+++ packages/ins/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/ins@2.0.3...@cosmos-kit/ins@2.1.0) (2023-07-28)
+
+**Note:** Version bump only for package @cosmos-kit/ins
+
+## [2.0.3](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/ins@2.0.2...@cosmos-kit/ins@2.0.3) (2023-07-28)
+
+**Note:** Version bump only for package @cosmos-kit/ins
+
 ## [2.0.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/ins@2.0.1...@cosmos-kit/ins@2.0.2) (2023-07-21)
 
 **Note:** Version bump only for package @cosmos-kit/ins