+++ packages/react/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react@2.1.0...@cosmos-kit/react@2.1.1) (2023-07-28)
+
+**Note:** Version bump only for package @cosmos-kit/react
+
+# [2.1.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react@2.0.5...@cosmos-kit/react@2.1.0) (2023-07-28)
+
+**Note:** Version bump only for package @cosmos-kit/react
+
+## [2.0.5](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react@2.0.4...@cosmos-kit/react@2.0.5) (2023-07-28)
+
+**Note:** Version bump only for package @cosmos-kit/react
+
+## [2.0.4](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react@2.0.3...@cosmos-kit/react@2.0.4) (2023-07-27)
+
+### Bug Fixes
+
+- **react:** undefined runtime error ([0abce01](https://github.com/cosmology-tech/cosmos-kit/commit/0abce01b677dfc5eb88961e339569000bb838b59))
+
 ## [2.0.3](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react@2.0.2...@cosmos-kit/react@2.0.3) (2023-07-24)
 
 **Note:** Version bump only for package @cosmos-kit/react